requirements-build.txt
# Build dependencies
pyinstaller==6.3.0
wheel==0.42.0
setuptools==69.0.3

# Runtime dependencies (also needed for building)
pygame==2.6.1
playsound==1.3.0
requests==2.32.3
Pillow==11.2.1
keyboard==0.13.5
openai==1.12.0
google-generativeai==0.8.3
sseclient-py==1.8.0
litellm==1.30.7
jaraco.text==3.10.0
speedtest-cli==2.1.3
urllib3==2.2.1
psutil==5.9.8